feat: make the math engine an opt-in, tree-shakeable entry point
The MathJax + STIX Two Math engine (~3 MB) was statically imported by the
docx and pptx renderer chunks, so every consumer of `@silurus/ooxml/docx`
or `/pptx` shipped the full ~3 MB in their initial bundle even when their
documents contained no equations (the runtime <script> injection only made
*execution* lazy, not the bytes).

Invert the dependency via injection: the renderers depend only on the light
`MathRenderer` contract (type + engine-free extents/recolor helpers). The
heavy engine moves to `packages/core/src/math/engine.ts` and is published as
a separate entry point — `@silurus/ooxml/math` — which exposes a single
named export `math`. Consumers opt in with a named import:

    import { DocxViewer } from '@silurus/ooxml/docx';
    import { math } from '@silurus/ooxml/math';
    new DocxViewer(canvas, { math });

Omit it and the engine is referenced nowhere, so a bundler drops the ~3 MB
entirely (equations are skipped). xlsx never references the engine. Works
for PptxViewer and the headless DocxDocument / PptxPresentation APIs too.

Verified on the build: the engine asset lives solely in `math.mjs`; no other
chunk imports it; docx/pptx pull only the light `mathjax` (0.45 kB) and
`mathml` (52 kB) chunks. Equation fidelity confirmed in Storybook on
sample-8 (inline italic n, ⇔, display coordinate tuples, ℝ² double-struck).

BREAKING CHANGE: equations in docx/pptx no longer render automatically.
Import { math } from '@silurus/ooxml/math' and pass it as the `math`
viewer/load option.

- core: split math/mathjax.ts (light) from math/engine.ts (heavy); add the
  MathRenderer contract; add `math` to RenderOptions; drop loadMathJax/
  mathMLToSvg from the package index.
- root: add `src/math.ts` entry (named `math` export), vite `math` lib
  entry, `./math` package export.
- docx/pptx: thread `math` through viewers, document/presentation, and
  prepareMathRuns/prepareSlideMath; skip math when no engine is provided.
- stories: wire the engine into the demo viewers so equation samples render.
- docs: README "Rendering equations" section + corrected bundle-size note;
  site api-reference `math` option on DocxViewer/DocxDocument/PptxViewer and
  PptxPresentation.renderSlide.
