+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@silurus/ooxml",
-  "version": "0.48.0",
+  "version": "0.48.1",
   "description": "Browser-based OOXML viewer (docx/xlsx/pptx) — Rust/WASM parser + Canvas renderer",
   "license": "MIT",
   "author": "Yuki Yokotani <silurus.dev@gmail.com>",
@@ -49,6 +49,10 @@
     "./pptx": {
       "import": "./dist/pptx.mjs",
       "types": "./dist/types/pptx.d.ts"
+    },
+    "./math": {
+      "import": "./dist/math.mjs",
+      "types": "./dist/types/math.d.ts"
     }
   },
   "packageManager": "pnpm@10.33.0+sha512.10568bb4a6afb58c9eb3630da90cc9516417abebd3fabbe6739f0ae795728da1491e9db5a544c76ad8eb7570f5c4bb3d6c637b2cb41bfdcdb47fa823c8649319",
