+++ packages/core/src/math/mathjax.ts
@@ -1,58 +1,10 @@
-// MathML → SVG via a pre-bundled MathJax v4 + STIX Two Math converter.
+// Light, dependency-free helpers for the math feature.
 //
-// The converter (engine + statically-baked STIX2 font, all glyph ranges) is
-// built by `build/build-mathjax.mjs` into the opaque asset
-// `assets/mathjax-stix2.js`. We load that asset *lazily* (only when a document
-// actually contains equations): the `new URL(...)` lives inside a function so
-// non-math viewers (xlsx) tree-shake it out, and keeping it pre-bundled stops
-// the consuming app's bundler from re-bundling — and over-including — the
-// MathJax source. The asset is self-contained: DOM-free internally, zero
-// network, zero cross-origin requests. It exposes `globalThis.__ooxmlStix2`.
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
-interface Stix2Engine {
-  /** MathML string → standalone `<svg>…</svg>` (currentColor fill, viewBox in 1000-units/em). */
-  mathml2svg(mathml: string): string;
-}
-
-let enginePromise: Promise<Stix2Engine> | null = null;
-
-function resolveAssetUrl(): string {
-  return new URL('../../assets/mathjax-stix2.js', import.meta.url).href;
-}
-
-function loadScript(src: string): Promise<void> {
-  return new Promise((resolve, reject) => {
-    const s = document.createElement('script');
-    s.src = src;
-    s.async = true;
-    s.onload = () => resolve();
-    s.onerror = () => reject(new Error(`Failed to load math engine from ${src}`));
-    document.head.appendChild(s);
-  });
-}
-
-function ensureEngine(): Promise<Stix2Engine> {
-  if (enginePromise) return enginePromise;
-  enginePromise = (async () => {
-    const existing = (globalThis as any).__ooxmlStix2 as Stix2Engine | undefined;
-    if (existing) return existing;
-    if (typeof document === 'undefined') {
-      throw new Error('Math rendering requires a DOM (browser environment)');
-    }
-    await loadScript(resolveAssetUrl());
-    const engine = (globalThis as any).__ooxmlStix2 as Stix2Engine | undefined;
-    if (!engine) throw new Error('Math engine failed to initialize');
-    return engine;
-  })();
-  return enginePromise;
-}
-
-/** Preload the math engine. Call once before rendering equations. */
-export async function loadMathJax(): Promise<void> {
-  await ensureEngine();
-}
+// IMPORTANT: this module must stay free of the heavy MathJax engine asset (that
+// lives in `./engine`, referenced via `new URL`). The docx/pptx renderers
+// import ONLY from here (extents math, recolor, the `MathRenderer` contract),
+// so the ~3 MB engine asset tree-shakes out of their bundles. The engine is
+// injected at runtime via the `math` viewer option — see `MathRenderer`.
 
 export interface MathSvg {
   /** standalone `<svg>…</svg>` markup. */
@@ -61,6 +13,27 @@ export interface MathSvg {
   widthEm: number;
   ascentEm: number;
   descentEm: number;
+}
+
+/**
+ * The math engine contract a viewer needs to render equations. Satisfied by the
+ * `math` named export of the separate `@silurus/ooxml/math` entry point, which
+ * the consumer opts into:
+ *
+ * ```ts
+ * import { DocxViewer } from '@silurus/ooxml/docx';
+ * import { math } from '@silurus/ooxml/math';
+ * new DocxViewer(canvas, { math });
+ * ```
+ *
+ * Omit it and the equation engine (MathJax + STIX Two Math, ~3 MB) is never
+ * imported, so a bundler drops it entirely.
+ */
+export interface MathRenderer {
+  /** Preload the engine. Called once before converting equations. */
+  loadMathJax(): Promise<void>;
+  /** MathML string → standalone SVG + baseline-relative em extents. */
+  mathMLToSvg(mathml: string): Promise<MathSvg>;
 }
 
 const UNITS_PER_EM = 1000;
@@ -79,13 +52,6 @@ export function svgExtents(svg: string): { widthEm: number; ascentEm: number; de
     ascentEm: -minY / UNITS_PER_EM,
     descentEm: (minY + h) / UNITS_PER_EM,
   };
-}
-
-/** Convert a MathML string to a standalone SVG + its baseline-relative extents. */
-export async function mathMLToSvg(mathml: string): Promise<MathSvg> {
-  const engine = await ensureEngine();
-  const svg = engine.mathml2svg(mathml);
-  return { svg, ...svgExtents(svg) };
 }
 
 /** Replace MathJax's `currentColor` placeholders with an explicit color (for raster). */
